chore: move `generic_argument` to `generic`
