+++ libs/@local/hashql/core/src/type/kind/generic.rs
