+++ libs/@local/hashql/core/src/type/kind/intersection.rs
@@ -645,7 +645,7 @@ mod test {
             },
             kind::{
                 OpaqueType, Param, TypeKind,
-                generic_argument::{GenericArgument, GenericArgumentId},
+                generic::{GenericArgument, GenericArgumentId},
                 infer::HoleId,
                 intrinsic::{DictType, IntrinsicType},
                 primitive::PrimitiveType,
